update parquet-file.jar from 0.0.0 to 0.0.2
