+++ dataSources/parquet/src/main/java/cn/edu/tsinghua/iginx/parquet/io/parquet/FilterUtils.java
@@ -24,9 +24,9 @@ import cn.edu.tsinghua.iginx.parquet.util.Constants;
 import cn.edu.tsinghua.iginx.utils.Pair;
 import java.util.Objects;
 import javax.annotation.Nonnull;
-import org.apache.parquet.filter2.predicate.FilterApi;
-import org.apache.parquet.filter2.predicate.FilterPredicate;
-import org.apache.parquet.filter2.predicate.Operators;
+import shaded.iginx.org.apache.parquet.filter2.predicate.FilterApi;
+import shaded.iginx.org.apache.parquet.filter2.predicate.FilterPredicate;
+import shaded.iginx.org.apache.parquet.filter2.predicate.Operators;
 
 class FilterUtils {
