+++ dataSources/parquet/src/main/java/cn/edu/tsinghua/iginx/parquet/io/parquet/IParquetWriter.java
@@ -23,14 +23,15 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.parquet.hadoop.metadata.ParquetMetadata;
-import org.apache.parquet.io.LocalOutputFile;
-import org.apache.parquet.io.OutputFile;
-import org.apache.parquet.local.ParquetFileWriter;
-import org.apache.parquet.local.ParquetRecordWriter;
-import org.apache.parquet.local.ParquetWriteOptions;
-import org.apache.parquet.schema.MessageType;
-import org.apache.parquet.schema.TypeUtil;
+import shaded.iginx.org.apache.parquet.ParquetWriteOptions;
+import shaded.iginx.org.apache.parquet.bytes.HeapByteBufferAllocator;
+import shaded.iginx.org.apache.parquet.hadoop.ParquetFileWriter;
+import shaded.iginx.org.apache.parquet.hadoop.ParquetRecordWriter;
+import shaded.iginx.org.apache.parquet.hadoop.metadata.ParquetMetadata;
+import shaded.iginx.org.apache.parquet.io.LocalOutputFile;
+import shaded.iginx.org.apache.parquet.io.OutputFile;
+import shaded.iginx.org.apache.parquet.schema.MessageType;
+import shaded.iginx.org.apache.parquet.schema.TypeUtil;
 
 public class IParquetWriter implements AutoCloseable {
 
@@ -44,7 +45,8 @@ public class IParquetWriter implements AutoCloseable {
   }
 
   public static Builder builder(Path path, MessageType schema) {
-    return new Builder(new LocalOutputFile(path), schema);
+    return new Builder(
+        new LocalOutputFile(path, new HeapByteBufferAllocator(), Integer.MAX_VALUE), schema);
   }
 
   public void write(IRecord record) throws IOException {
@@ -57,12 +59,8 @@ public class IParquetWriter implements AutoCloseable {
   }
 
   public ParquetMetadata flush() throws IOException {
-    try {
-      internalWriter.close();
-      return fileWriter.getFooter();
-    } catch (InterruptedException e) {
-      throw new IOException(e);
-    }
+    internalWriter.close();
+    return fileWriter.getFooter();
   }
 
   public static class Builder {
@@ -85,8 +83,7 @@ public class IParquetWriter implements AutoCloseable {
       TypeUtil.checkValidWriteSchema(schema);
       ParquetFileWriter fileWriter = new ParquetFileWriter(outputFile, options);
       ParquetRecordWriter<IRecord> recordWriter =
-          new ParquetRecordWriter<>(
-              fileWriter, new IRecordDematerializer(schema), schema, extraMetaData, options);
+          new ParquetRecordWriter<>(fileWriter, new IRecordDematerializer(schema), options);
       return new IParquetWriter(recordWriter, fileWriter);
     }
 
@@ -101,7 +98,7 @@ public class IParquetWriter implements AutoCloseable {
     }
 
     public Builder withPageSize(int pageSize) {
-      optionsBuilder.withPageSize(pageSize);
+      optionsBuilder.asParquetPropertiesBuilder().withPageSize(pageSize);
       return this;
     }
   }
