+++ dataSources/parquet/src/main/java/cn/edu/tsinghua/iginx/parquet/io/parquet/IRecordDematerializer.java
@@ -16,14 +16,15 @@
 
 package cn.edu.tsinghua.iginx.parquet.io.parquet;
 
+import java.util.Collections;
 import java.util.Map;
-import org.apache.parquet.io.api.Binary;
-import org.apache.parquet.io.api.RecordConsumer;
-import org.apache.parquet.io.api.RecordDematerializer;
-import org.apache.parquet.schema.GroupType;
-import org.apache.parquet.schema.MessageType;
-import org.apache.parquet.schema.PrimitiveType;
-import org.apache.parquet.schema.Type;
+import shaded.iginx.org.apache.parquet.io.api.Binary;
+import shaded.iginx.org.apache.parquet.io.api.RecordConsumer;
+import shaded.iginx.org.apache.parquet.io.api.RecordDematerializer;
+import shaded.iginx.org.apache.parquet.schema.GroupType;
+import shaded.iginx.org.apache.parquet.schema.MessageType;
+import shaded.iginx.org.apache.parquet.schema.PrimitiveType;
+import shaded.iginx.org.apache.parquet.schema.Type;
 
 class IRecordDematerializer extends RecordDematerializer<IRecord> {
 
@@ -47,6 +48,16 @@ class IRecordDematerializer extends RecordDematerializer<IRecord> {
       addGroup(schema, record);
       recordConsumer.endMessage();
     }
+  }
+
+  @Override
+  public MessageType getSchema() {
+    return schema;
+  }
+
+  @Override
+  public Map<String, String> getExtraMetaData() {
+    return Collections.emptyMap();
   }
 
   private void addGroup(GroupType groupType, IRecord record) {
